job_portal2/requirements.txt
Django==2.0.1
django-extensions==2.1.2
django-filter==2.0.0
django-mysql==2.4.1
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.13
pytz==2018.5
six==1.11.0
